fix: 배송 요청 프론트 입력란 백에서 받을 수 있는 형식으로 wrap, 전송 구조 수정

+++ backend/squirrelLogistics/src/main/java/com/gpt/squirrelLogistics/common/LatLng.java
@@ -1,11 +1,13 @@
 package com.gpt.squirrelLogistics.common;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
 @Data
 @AllArgsConstructor
 public class LatLng {
-    private double lat;
-    private double lng;
+	private BigDecimal lat;
+	private BigDecimal lng;
 }
